Kernel: Make FS::block_size a size_t

+++ Kernel/FileSystem/FileSystem.cpp
@@ -107,7 +107,7 @@ void FS::lock_all()
     }
 }
 
-void FS::set_block_size(int block_size)
+void FS::set_block_size(size_t block_size)
 {
     ASSERT(block_size > 0);
     if (block_size == m_block_size)
